Added more params

+++ learning.py
@@ -12,7 +12,7 @@ def main(EPOCH=10,batch_size=32):
     x,y=dataset[0],dataset[1]
     y=tf.keras.layers.Rescaling(1.0/y.max())(y)
     x=np.array(x,dtype="uint8")
-    x_train,x_test,y_train,y_test=train_test_split(x,np.array(y).reshape(y.shape[0],64),test_size=0.2,random_state=0)
+    x_train,x_test,y_train,y_test=train_test_split(x,np.array(y,dtype="float32").reshape(y.shape[0],64),test_size=0.2,random_state=0)
     #=train_test_split(,random_state=0)
     test_ds = tf.data.Dataset.from_tensor_slices(
                 (x_test,y_test)
